refactor: utilize layout wrappers for UI elements and providers
