+++ theme/gatsby-browser.js
@@ -1,4 +1,6 @@
 import './src/styles/global.css'
+import 'prism-themes/themes/prism-material-light.css'
+import 'prismjs/plugins/line-numbers/prism-line-numbers.css'
 
-require('prism-themes/themes/prism-material-light.css')
-require('prismjs/plugins/line-numbers/prism-line-numbers.css')
+export { default as wrapPageElement } from './wrapPageElement'
+export { default as wrapRootElement } from './wrapRootElement'
