+++ theme/gatsby-ssr.js
@@ -1,0 +1,2 @@
+export { default as wrapPageElement } from './wrapPageElement'
+export { default as wrapRootElement } from './wrapRootElement'
