+++ theme/src/components/layout.js
@@ -1,18 +1,9 @@
 /** @jsx jsx */
+import { jsx } from 'theme-ui'
 import { Fragment } from 'react'
-import { jsx, ThemeProvider } from 'theme-ui'
-import { MDXProvider } from '@mdx-js/react'
 
-import theme from '../gatsby-plugin-theme-ui'
+import Footer from './footer'
 import TopNavigation from './top-navigation'
-import Emoji from '../shortcodes/emoji'
-import YouTube from '../shortcodes/youtube'
-
-const components = {
-  pre: ({ children }) => <Fragment>{children}</Fragment>
-}
-
-const shortcodes = { Emoji, YouTube }
 
 /**
  * Layout
@@ -21,13 +12,13 @@ const shortcodes = { Emoji, YouTube }
  * the default navigation, theme styles, and any important providers. Use shadowing
  * to extend this component and attach additional contexts and providers.
  */
-const Layout = ({ children, hideNavigation }) => (
-  <ThemeProvider theme={theme} components={components}>
-    <MDXProvider components={shortcodes}>
-      {!hideNavigation && <TopNavigation />}
-      {children}
-    </MDXProvider>
-  </ThemeProvider>
+const Layout = ({ children, path }) => (
+  <Fragment>
+    {/* NOTE(chrisvogt): hide the top navigation on the home page */}
+    {path !== '/' && <TopNavigation />}
+    {children}
+    <Footer />
+  </Fragment>
 )
 
 export default Layout
