+++ theme/src/pages/latest.js
@@ -1,18 +1,17 @@
 /** @jsx jsx */
 import { Container, jsx, Styled } from 'theme-ui'
 import { Flex } from '@theme-ui/components'
+import { Fragment } from 'react'
 import { graphql } from 'gatsby'
 
 import { getPosts } from '../hooks/use-recent-posts'
-import Footer from '../components/footer'
-import Layout from '../components/layout'
 import PostCard from '../components/widgets/recent-posts/post-card'
 import SEO from '../components/seo'
 
 export default ({ data }) => {
   const posts = getPosts(data)
   return (
-    <Layout>
+    <Fragment>
       <SEO
         title='Latest Content'
         description='A list of the most recent articles published on my blog.'
@@ -51,9 +50,7 @@ export default ({ data }) => {
           </Styled.div>
         </Container>
       </Flex>
-
-      <Footer />
-    </Layout>
+    </Fragment>
   )
 }
