+++ theme/src/styles/global.css
@@ -2,7 +2,7 @@ html,
 body,
 #___gatsby,
 #gatsby-focus-wrapper {
-  height: 100%;
+  min-height: 100%;
 }
 
 #gatsby-focus-wrapper {
