+++ theme/src/templates/media.js
@@ -1,5 +1,6 @@
 /** @jsx jsx */
 import { Container, Flex, jsx, Styled, useThemeUI } from 'theme-ui'
+import { Fragment } from 'react'
 import { graphql } from 'gatsby'
 import { Heading } from '@theme-ui/components'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
@@ -7,8 +8,6 @@ import PropTypes from 'prop-types'
 
 import isDarkMode from '../helpers/isDarkMode'
 
-import Footer from '../components/footer'
-import Layout from '../components/layout'
 import SEO from '../components/seo'
 
 import SoundCloud from '../shortcodes/soundcloud'
@@ -28,7 +27,7 @@ const MediaTemplate = ({ data }) => {
   const soundcloudId = mdx.frontmatter.soundcloudId
 
   return (
-    <Layout>
+    <Fragment>
       <SEO
         article={true}
         description={description}
@@ -73,9 +72,7 @@ const MediaTemplate = ({ data }) => {
           </div>
         </Container>
       </Flex>
-
-      <Footer />
-    </Layout>
+    </Fragment>
   )
 }
