+++ theme/src/templates/post.js
@@ -1,12 +1,11 @@
 /** @jsx jsx */
 import { Container, Flex, jsx, Styled } from 'theme-ui'
+import { Fragment } from 'react'
 import { graphql } from 'gatsby'
 import { Heading } from '@theme-ui/components'
 import { MDXRenderer } from 'gatsby-plugin-mdx'
 import PropTypes from 'prop-types'
 
-import Footer from '../components/footer'
-import Layout from '../components/layout'
 import SEO from '../components/seo'
 
 const PostTemplate = ({ data }) => {
@@ -19,7 +18,7 @@ const PostTemplate = ({ data }) => {
   const title = mdx.frontmatter.title
 
   return (
-    <Layout>
+    <Fragment>
       <SEO
         article={true}
         description={description}
@@ -46,9 +45,7 @@ const PostTemplate = ({ data }) => {
           </div>
         </Container>
       </Flex>
-
-      <Footer />
-    </Layout>
+    </Fragment>
   )
 }
